# backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Alpine Ecommerce API",
    "version": "1.0.0",
    "description": "# Alpine Ecommerce Backend API\n\nComplete API documentation for Alpine Ecommerce e-commerce platform.\n\n## Features\n- 🔐 JWT Authentication\n- 📦 Product Management\n- 🛒 Order Processing\n- 💳 Payment Integration (Razorpay)\n- 👥 Customer Management\n- 📊 Dashboard & Reports\n- 📝 Error Logging\n\n## Authentication\nMost endpoints require JWT authentication. Include the token in the Authorization header:\n```\nAuthorization: Bearer <your-token>\n```\n\n## Base URL\n- **Current Server**: http://13.235.53.15:8081/api/v1\n- **Configuration**: \n  - Set `API_BASE_URL` in .env for full control (recommended)\n  - Or set `PRODUCTION_URL` for production domain\n  - Or set `DOMAIN`, `PORT`, and `PROTOCOL` for auto-construction\n\n## Response Format\nAll responses follow a consistent format:\n```json\n{\n  \"responseStatusCode\": 200,\n  \"isError\": false,\n  \"errorData\": null,\n  \"successData\": { ... }\n}\n```\n\n## Error Handling\nErrors are returned with appropriate status codes:\n```json\n{\n  \"responseStatusCode\": 400,\n  \"isError\": true,\n  \"errorData\": {\n    \"displayMessage\": \"Error description\",\n    \"apiErrorType\": 1\n  }\n}\n```",
    "contact": {
      "name": "Alpine Ecommerce API Support",
      "email": "support@wildvalleyfoods.in"
    },
    "license": {
      "name": "Proprietary"
    },
    "servers": [
      {
...
      }
    ],
...
        }
...
  },
  "paths": {
...
    "/api/v1/license": {
      "post": {
        "tags": [
          "Licenses"
        ],
        "summary": "Create license (Admin only)",
        "description": "Creates a new license",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "reqData"
                ],
                "properties": {
                  "reqData": {
                    "type": "object",
                    "required": [
                      "name"
                    ],
                    "properties": {
                      "name": {
                        "type": "string",
                        "example": "Premium License"
                      },
                      "description": {
                        "type": "string",
                        "example": "License description"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "License created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "get": {
        "tags": [
          "Licenses"
        ],
        "summary": "Get all licenses",
        "description": "Returns all licenses",
        "security": [],
        "responses": {
          "200": {
            "description": "Licenses retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/license/{id}": {
      "get": {
        "tags": [
          "Licenses"
        ],
        "summary": "Get license by ID",
        "description": "Returns a specific license by its ID",
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "License ID"
          }
        ],
        "responses": {
          "200": {
            "description": "License retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "tags": [
          "Licenses"
        ],
        "summary": "Update license (Admin only)",
        "description": "Updates an existing license",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "License ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "reqData"
                ],
                "properties": {
                  "reqData": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "example": "Updated License Name"
                      },
                      "description": {
                        "type": "string",
                        "example": "Updated description"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "License updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "Licenses"
        ],
        "summary": "Delete license (Admin only)",
        "description": "Deletes a license by ID",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "License ID"
          }
        ],
        "responses": {
          "200": {
            "description": "License deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /register or /login endpoint. Enter your token (with or without \"Bearer \" prefix)"
      }
    }
  },
...
}